server: swap active and passive degrees if needed when creating Query
If active degree < passive degree, swap the degrees and other
related properties when constructing ProblemProps objcet.
The object, among other things, is used when constructing
Query object. Thus, now when a user queries for problems
whose active and passive degrees are 2 and 3, respectively,
internally the query will look for problems with degrees 3 and 2
instead (swapped). This is because problems are always stored so that
active degree is >= than passive degree.